More fixes.
It was previously impossible to specify popups should be used if you had previously set the keyguard to use redirects for all function calls, since the only way to do it was to specify `redirectBehavior` as `null` but that's a falsy value so it gets ignored as if it hadn't been declared.

Rather than change how I check the `options` objects to use hasOwnProperty or something else, I just made it so you would use `redirectBehavior : { popup : true }` which I think is more clear  than `redirectBehavior : null` or `redirectBehavior: { }`.